fix(prhls): テストが動作するよう修正
- concurrentlyを削除。
- ts-nodeを追加、テスト実行はこちらで。
- diffを追加、差分表示用。
- Handlerクラスをlib/handler.tsに分離しlib/index.tsからインポート。
- テスト用のcreateConnection()に入出力ストリームを明示して渡す。
- 出力ストリームはTransform。
  - スナップショットを記録する場合のみファイルストリームにpipeする。
  - サーバーからのレスポンスだけを、pipeされたストリームと標準出力へ流す。
- exitイベントハンドラーでレスポンスを比較。
  - レスポンスはassert()と===で雑に比較。
  - 一致しないときはunified diffを表示。
