+++ package.json
@@ -17,7 +17,8 @@
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
     "vscode:update": "npm_package_engines_vscode='^1.11.0' node ./node_modules/vscode/bin/install",
     "build": "tsfmt -r && tsc",
-    "test": "concurrently --kill-others \"node lib/index.js --stdio\" \"node test/input.js\""
+    "test": "ts-node test/input.ts",
+    "test:updateSnapshot": "UPDATE_SNAPSHOT=1 ts-node test/input.ts"
   },
   "dependencies": {
     "prh": "^5.3.0",
@@ -25,12 +26,14 @@
     "vscode-uri": "2.1.1"
   },
   "devDependencies": {
+    "@types/diff": "^4.0.2",
     "@types/mocha": "^2.2.32",
     "@types/node": "^12.12.0",
     "@types/vscode": "^1.40.0",
-    "concurrently": "^5.0.0",
     "conventional-changelog-cli": "^1.3.2",
+    "diff": "^4.0.1",
     "mocha": "^3.5.0",
+    "ts-node": "^8.5.4",
     "typescript": "3.6.3",
     "typescript-formatter": "^6.0.0",
     "vscode-jsonrpc": "4.0.0",
